move state up, reduce conditional logic in composed chart handler
